refactor: Adjustments to folders and namespaces
