+++ Backend/CampusTradeSystem/CampusTrade.API/Controllers/NotificationController.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using CampusTrade.API.Services.Auth;
-using CampusTrade.API.Utils.Notificate;
+using CampusTrade.API.Infrastructure.Utils.Notificate;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
