+++ Backend/CampusTradeSystem/CampusTrade.API/Program.cs
@@ -1,8 +1,8 @@
 using System.Text;
 using System.Text.Encodings.Web;
 using CampusTrade.API.Data;
-using CampusTrade.API.Extensions;
-using CampusTrade.API.Middleware;
+using CampusTrade.API.Infrastructure.Extensions;
+using CampusTrade.API.Infrastructure.Middleware;
 using CampusTrade.API.Options;
 using CampusTrade.API.Services.Background;
 using CampusTrade.API.Services.Cache;
@@ -291,7 +291,7 @@ try
     app.MapControllers();
 
     // 映射SignalR Hub
-    app.MapHub<CampusTrade.API.Models.Hubs.NotificationHub>("/api/notification-hub");
+    app.MapHub<CampusTrade.API.Infrastructure.Hubs.NotificationHub>("/api/notification-hub");
 
     Log.Information("Campus Trade API 启动完成");
