+++ Backend/CampusTradeSystem/CampusTrade.API/Services/Auth/AuthService.cs
@@ -2,8 +2,8 @@ using CampusTrade.API.Models.DTOs.Auth;
 using CampusTrade.API.Models.Entities;
 using CampusTrade.API.Repositories.Interfaces;
 using CampusTrade.API.Services.Interfaces;
-using CampusTrade.API.Utils;
-using CampusTrade.API.Utils.Security;
+using CampusTrade.API.Infrastructure.Utils;
+using CampusTrade.API.Infrastructure.Utils.Security;
 using Microsoft.EntityFrameworkCore;
 using Serilog;
