+++ Backend/CampusTradeSystem/CampusTrade.API/Services/Cache/ProductCacheService.cs
@@ -8,7 +8,7 @@ using CampusTrade.API.Models.Entities;
 using CampusTrade.API.Options;
 using CampusTrade.API.Services.Cache;
 using CampusTrade.API.Services.Interfaces;
-using CampusTrade.API.Utils.Cache;
+using CampusTrade.API.Infrastructure.Utils.Cache;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
