+++ Backend/CampusTradeSystem/CampusTrade.API/Services/Notification/NotifiSenderService.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using CampusTrade.API.Data;
 using CampusTrade.API.Models.Entities;
-using CampusTrade.API.Models.Hubs;
+using CampusTrade.API.Infrastructure.Hubs;
 using CampusTrade.API.Services.Email;
-using CampusTrade.API.Utils.Notificate;
+using CampusTrade.API.Infrastructure.Utils.Notificate;
 using Microsoft.AspNetCore.SignalR;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
