+++ Backend/CampusTradeSystem/CampusTrade.API/Services/Notification/NotifiService.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using CampusTrade.API.Data;
 using CampusTrade.API.Models.Entities;
 using CampusTrade.API.Services.Background;
-using CampusTrade.API.Utils.Notificate;
+using CampusTrade.API.Infrastructure.Utils.Notificate;
 using Microsoft.EntityFrameworkCore;
 
 namespace CampusTrade.API.Services.Auth
