+++ Backend/CampusTradeSystem/CampusTrade.API/infrastructure/Middleware/SecurityHeadersMiddleware.cs
@@ -1,0 +1,44 @@
+namespace CampusTrade.API.Infrastructure.Middleware
+{
+    /// <summary>
+    /// 安全头中间件
+    /// 添加必要的安全响应头
+    /// </summary>
+    public class SecurityHeadersMiddleware
+    {
+        private readonly RequestDelegate _next;
+        private readonly ILogger<SecurityHeadersMiddleware> _logger;
+
+        public SecurityHeadersMiddleware(RequestDelegate next, ILogger<SecurityHeadersMiddleware> logger)
+        {
+            _next = next;
+            _logger = logger;
+        }
+
+        public async Task InvokeAsync(HttpContext context)
+        {
+            // 添加安全响应头
+            context.Response.Headers.Append("X-Content-Type-Options", "nosniff");
+            context.Response.Headers.Append("X-Frame-Options", "DENY");
+            context.Response.Headers.Append("X-XSS-Protection", "1; mode=block");
+            context.Response.Headers.Append("Referrer-Policy", "strict-origin-when-cross-origin");
+            context.Response.Headers.Append("Content-Security-Policy", "default-src 'self'");
+
+            // 删除服务器信息头
+            context.Response.Headers.Remove("Server");
+
+            await _next(context);
+        }
+    }
+
+    /// <summary>
+    /// 安全头中间件扩展
+    /// </summary>
+    public static class SecurityHeadersMiddlewareExtensions
+    {
+        public static IApplicationBuilder UseSecurityHeaders(this IApplicationBuilder builder)
+        {
+            return builder.UseMiddleware<SecurityHeadersMiddleware>();
+        }
+    }
+}
