+++ Backend/CampusTradeSystem/CampusTrade.Tests/UnitTests/Middleware/SecurityMiddlewareTests.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using CampusTrade.API.Middleware;
+using CampusTrade.API.Infrastructure.Middleware;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
