+++ Backend/CampusTradeSystem/CampusTrade.Tests/UnitTests/Services/NotificationServiceTests.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using CampusTrade.API.Data;
 using CampusTrade.API.Models.Entities;
 using CampusTrade.API.Services.Auth;
-using CampusTrade.API.Utils.Notificate;
+using CampusTrade.API.Infrastructure.Utils.Notificate;
 using CampusTrade.Tests.Helpers;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
